// src/types/UserData.ts
// Generated by https://quicktype.io

export interface UserData {
  status: string;
  message: string;
  data: Data;
  code: number;
}

export interface Data {
  userData: UserDataClass;
}

export interface UserDataClass {
  id: number;
  userLname: string;
  userFname: string;
  userMname: string;
  userEmail: string;
  userContact: string;
  userDob: string;
  userCustomerNumber: string;
  createdAt: string;
  updatedAt: string;
  address: any[];
}